Wired up the progress bar so that it actually leverages the number of keys loaded vs the total amount of keys to display the progress.

+++ src/GuiR/Redis/KeyInfo.cs
@@ -10,9 +10,9 @@ namespace GuiR.Redis
 
         public ConnectionMultiplexer ConnectionMultiplexer { get; }
 
-        public int KeyCount { get; }
+        public long KeyCount { get; }
 
-        public KeyInfo(IEnumerable<string> keys, ConnectionMultiplexer connectionMultiplexer, int keyCount)
+        public KeyInfo(IEnumerable<string> keys, ConnectionMultiplexer connectionMultiplexer, long keyCount)
         {
             Keys = keys;
             ConnectionMultiplexer = connectionMultiplexer;
